Cloud data initialization is only done by CloudManipulator instead of CloudSimulator.
